Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -13,7 +13,7 @@ int main() {
         printf("%d",a-b);
     }
     else if(c=='/'){
-        float m=a/b;
+        int m=a/b;
         // printf("%f",m);
     }
     else{
